last module to process videos and dirs, move helper there
